# src/swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Digital Department API",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
    "/download/{filePath}/{fileName}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "filePath",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}